Move email templates to views
