+++ app/services/NotificationService.scala
@@ -3,12 +3,10 @@ package services
 import akka.stream.{ActorAttributes, Materializer, Supervision}
 import akka.stream.scaladsl.{RestartSource, Sink, Source}
 import constants.Constants
-import java.time.ZonedDateTime
 import java.util.UUID
 import javax.inject.{Inject, Singleton}
 import controllers.routes
 import helper.EmailHelper.quoteEmailPhrase
-import helper.Time
 import models._
 import models.mandat.Mandat
 import play.api.Logger
@@ -17,14 +15,7 @@ import play.api.libs.mailer.MailerClient
 import play.api.libs.mailer.Email
 import scala.concurrent.{ExecutionContext, Future}
 import scala.concurrent.duration._
-
-object NotificationService {
-
-  val magicLinkEmailSubject = "Connexion sécurisée à Administration+"
-
-  case class WeeklyEmailInfos(user: User, applicationsThatShouldBeClosed: List[Application])
-
-}
+import views.emails.{common, WeeklyEmailInfos}
 
 @Singleton
 class NotificationService @Inject() (
@@ -37,7 +28,6 @@ class NotificationService @Inject() (
     materializerProvider: MaterializerProvider,
     userService: UserService
 )(implicit executionContext: ExecutionContext) {
-  import NotificationService._
 
   implicit val materializer: Materializer = materializerProvider.get
 
@@ -172,15 +162,49 @@ class NotificationService @Inject() (
     sendMail(generateWelcomeEmail(newUser))
 
   def newMagicLinkEmail(user: User, loginToken: LoginToken, pathToRedirectTo: String) = {
-    val email = generateMagicLinkEmail(user, loginToken, pathToRedirectTo)
+    val absoluteUrl: String =
+      routes.LoginController.magicLinkAntiConsumptionPage().absoluteURL(https, host)
+    val bodyInner = common.magicLinkBody(
+      user,
+      loginToken,
+      absoluteUrl,
+      pathToRedirectTo,
+      tokenExpirationInMinutes
+    )
+    val email = Email(
+      subject = common.magicLinkSubject,
+      from = from,
+      to = List(s"${quoteEmailPhrase(user.name)} <${user.email}>"),
+      bodyHtml = Some(common.renderEmail(bodyInner))
+    )
     sendMail(email)
   }
 
-  def mandatSmsSent(mandatId: Mandat.Id, user: User): Unit =
-    sendMail(generateMandatSmsSentEmail(mandatId, user))
+  def mandatSmsSent(mandatId: Mandat.Id, user: User): Unit = {
+    val absoluteUrl: String =
+      routes.MandatController.mandat(mandatId.underlying).absoluteURL(https, host)
+    val bodyInner = common.mandatSmsSentBody(absoluteUrl)
+    val email = Email(
+      subject = common.mandatSmsSentSubject,
+      from = from,
+      to = List(s"${quoteEmailPhrase(user.name)} <${user.email}>"),
+      bodyHtml = Some(common.renderEmail(bodyInner))
+    )
+    sendMail(email)
+  }
 
-  def mandatSmsClosed(mandatId: Mandat.Id, user: User): Unit =
-    sendMail(generateMandatSmsClosedEmail(mandatId, user))
+  def mandatSmsClosed(mandatId: Mandat.Id, user: User): Unit = {
+    val absoluteUrl: String =
+      routes.MandatController.mandat(mandatId.underlying).absoluteURL(https, host)
+    val bodyInner = common.mandatSmsSentBody(absoluteUrl)
+    val email = Email(
+      subject = common.mandatSmsClosedSubject,
+      from = from,
+      to = List(s"${quoteEmailPhrase(user.name)} <${user.email}>"),
+      bodyHtml = Some(common.renderEmail(bodyInner))
+    )
+    sendMail(email)
+  }
 
   private def generateNotificationBALEmail(
       application: Application,
@@ -221,224 +245,29 @@ class NotificationService @Inject() (
 
   private def generateInvitationEmail(application: Application, answer: Option[Answer] = None)(
       invitedUser: User
-  ): Email =
+  ): Email = {
+    val absoluteUrl =
+      routes.ApplicationController.show(application.id).absoluteURL(https, host)
+    val bodyInner = common.invitationBody(application, answer, invitedUser, absoluteUrl)
     Email(
       subject = s"[A+] Nouvelle demande d'aide : ${application.subject}",
       from = from,
       to = List(s"${quoteEmailPhrase(invitedUser.name)} <${invitedUser.email}>"),
-      bodyHtml = Some(renderEmail(invitationEmailBody(application, answer, invitedUser)))
+      bodyHtml = Some(common.renderEmail(bodyInner))
     )
+  }
 
-  private def generateAnswerEmail(application: Application, answer: Answer)(user: User): Email =
+  private def generateAnswerEmail(application: Application, answer: Answer)(user: User): Email = {
+    val absoluteUrl =
+      routes.ApplicationController.show(application.id).absoluteURL(https, host)
+    val bodyInner = common.answerBody(application, answer, user, absoluteUrl)
     Email(
       subject = s"[A+] Nouvelle réponse pour : ${application.subject}",
       from = from,
       to = List(s"${quoteEmailPhrase(user.name)} <${user.email}>"),
-      bodyHtml = Some(renderEmail(answerEmailBody(application, answer, user)))
-    )
-
-  import scalatags.Text.all._
-
-  private def generateMagicLinkEmail(
-      user: User,
-      loginToken: LoginToken,
-      pathToRedirectTo: String
-  ): Email = {
-    val absoluteUrl: String =
-      routes.LoginController.magicLinkAntiConsumptionPage().absoluteURL(https, host)
-    val url = s"${absoluteUrl}?token=${loginToken.token}&path=$pathToRedirectTo"
-    val bodyInner =
-      List[Modifier](
-        s"Bonjour ${user.name},",
-        br,
-        br,
-        p(
-          "Vous pouvez maintenant accéder au service Administration+ en cliquant sur le lien suivant :",
-          br,
-          a(href := url, url),
-          br,
-          br,
-          "Ce lien est à ",
-          b("usage unique"),
-          " et est ",
-          b(s"valide $tokenExpirationInMinutes minutes"),
-          ", donc à utiliser avant ",
-          ZonedDateTime
-            .now(user.timeZone)
-            .plusMinutes(tokenExpirationInMinutes.toLong)
-            .format(Time.hourAndMinutesFormatter),
-          ".",
-          br,
-          br,
-          "Si vous avez des questions ou vous rencontrez un problème, ",
-          "n'hésitez pas à nous contacter sur ",
-          a(href := s"mailto:${Constants.supportEmail}", Constants.supportEmail),
-          br,
-          "Equipe Administration+"
-        )
-      ) ::: applicationEmailFooter
-
-    Email(
-      subject = magicLinkEmailSubject,
-      from = from,
-      to = List(s"${quoteEmailPhrase(user.name)} <${user.email}>"),
-      bodyHtml = Some(renderEmail(bodyInner))
+      bodyHtml = Some(common.renderEmail(bodyInner))
     )
   }
-
-  private def generateMandatSmsSentEmail(
-      mandatId: Mandat.Id,
-      user: User
-  ): Email = {
-    val url: String =
-      routes.MandatController.mandat(mandatId.underlying).absoluteURL(https, host)
-    val subject = s"[A+] Mandat initié par SMS"
-    val bodyInner =
-      List[Modifier](
-        span(
-          "Le SMS a bien été envoyé ! Vous recevrez un e-mail dès que l’usager aura répondu. ",
-          "Vous pouvez suivre l’échange en cliquant sur ce lien : ",
-          a(href := url, url)
-        )
-      ) ::: applicationEmailFooter
-    Email(
-      subject = subject,
-      from = from,
-      to = List(s"${quoteEmailPhrase(user.name)} <${user.email}>"),
-      bodyHtml = Some(renderEmail(bodyInner))
-    )
-  }
-
-  private def generateMandatSmsClosedEmail(
-      mandatId: Mandat.Id,
-      user: User
-  ): Email = {
-    val url: String =
-      routes.MandatController.mandat(mandatId.underlying).absoluteURL(https, host)
-    val subject = s"[A+] Vous avez reçu une réponse par SMS à votre mandat"
-    val bodyInner =
-      List[Modifier](
-        span(
-          "L’usager a répondu à votre demande de mandat. ",
-          "L’échange est disponible en cliquant sur ce lien : ",
-          a(href := url, url)
-        )
-      ) ::: applicationEmailFooter
-    Email(
-      subject = subject,
-      from = from,
-      to = List(s"${quoteEmailPhrase(user.name)} <${user.email}>"),
-      bodyHtml = Some(renderEmail(bodyInner))
-    )
-  }
-
-  private def applicationEmailFooter: List[Modifier] =
-    List(
-      br,
-      br,
-      b("Ne transférez pas cet email et n’y répondez pas directement."),
-      br,
-      i(
-        "- Vous pouvez transférer la demande à un autre utilisateur en ouvrant le lien ci-dessus",
-        br,
-        "- Si vous avez un problème ou besoin d’aide à propos de l’outil Administration+, contactez-nous sur ",
-        a(
-          href := s"mailto:${Constants.supportEmail}",
-          s"${Constants.supportEmail}"
-        ),
-        br,
-        "- Le navigateur Internet Explorer peut rencontrer des difficultés à accéder au site. Microsoft conseille depuis février 2019 de ne plus utiliser son navigateur historique qui n’est plus mis à jour depuis la sortie de Edge en 2015 et ne supporte donc plus les standards actuels du Web. ",
-        a(
-          href := "https://docs.aplus.beta.gouv.fr/faq/pourquoi-ne-plus-utiliser-le-navigateur-internet-explorer-de-microsoft",
-          "Pour en savoir plus"
-        )
-      )
-    )
-
-  private def invitationEmailBody(
-      application: Application,
-      answer: Option[Answer],
-      invitedUser: User
-  ): List[Modifier] = {
-    val url =
-      s"${routes.ApplicationController.show(application.id).absoluteURL(https, host)}?key=${invitedUser.key}"
-    List[Modifier](
-      s"Bonjour ${invitedUser.name},",
-      br,
-      br,
-      p(
-        answer.map(_.creatorUserName).getOrElse[String](application.creatorUserName),
-        " a besoin de vous.",
-        br,
-        "Cette personne vous a invité sur la demande suivante : ",
-        application.subject,
-        br,
-        "Vous pouvez voir la demande et y répondre en suivant ce lien : ",
-        br,
-        a(
-          href := url,
-          url
-        ),
-        br
-      )
-    ) ::: applicationEmailFooter
-  }
-
-  private def answerEmailBody(
-      application: Application,
-      answer: Answer,
-      user: User
-  ): List[Modifier] = {
-    val url =
-      s"${routes.ApplicationController.show(application.id).absoluteURL(https, host)}?key=${user.key}#answer-${answer.id}"
-    val irrelevantDefaultText: List[Modifier] =
-      if (answer.declareApplicationHasIrrelevant)
-        List[Modifier](
-          answer.creatorUserName.split('(').headOption.getOrElse[String]("L'agent"),
-          " a indiqué qu'",
-          b("il existe une procédure standard que vous pouvez utiliser pour cette demande"),
-          " vous aurez plus de détails dans sa réponse.",
-          br,
-          br
-        )
-      else
-        Nil
-    List[Modifier](
-      s"Bonjour ${user.name},",
-      br,
-      br,
-      p(
-        answer.creatorUserName,
-        " a donné une réponse sur la demande : ",
-        application.subject,
-        br,
-        irrelevantDefaultText,
-        "Vous pouvez consulter la réponse, y répondre ou la clôturer en suivant le lien suivant : ",
-        br,
-        a(
-          href := url,
-          url
-        )
-      )
-    ) ::: applicationEmailFooter
-  }
-
-  private def renderEmail(inner: List[Modifier]): String =
-    "<!DOCTYPE html>" + html(
-      head(
-        meta(
-          name := "viewport",
-          content := "width=device-width"
-        ),
-        meta(
-          httpEquiv := "Content-Type",
-          content := "text/html; charset=UTF-8"
-        )
-      ),
-      body(
-        inner: _*
-      )
-    )
 
   def weeklyEmails(): Future[Unit] =
     Source
@@ -467,69 +296,16 @@ class NotificationService @Inject() (
       }
 
   private def sendWeeklyEmail(infos: WeeklyEmailInfos): Future[Unit] = {
-    val subject = "[A+] Votre récapitulatif hebdomadaire"
-    val datePattern = "EEEE dd MMMM YYYY HH'h'mm"
-    val bodyInner =
-      List[Modifier](
-        s"Bonjour ${infos.user.name},",
-        br,
-        br,
-        p(
-          s"Vous avez ${infos.applicationsThatShouldBeClosed.size} ",
-          (
-            if (infos.applicationsThatShouldBeClosed.size <= 1)
-              "demande qui a reçu une réponse il y a "
-            else
-              "demandes qui ont reçu une réponse il y a ",
-          ),
-          s"plus de $daySinceLastAgentAnswerForApplicationsThatShouldBeClosed jours. ",
-          "Si votre échange est terminé, n’hésitez pas à ",
-          (
-            if (infos.applicationsThatShouldBeClosed.size <= 1)
-              "la clore, "
-            else
-              "les clore, "
-          ),
-          """en appuyant sur le bouton « Clore l’échange »."""
-        ),
-        ul(
-          infos.applicationsThatShouldBeClosed.map(application =>
-            li(
-              a(
-                href := routes.ApplicationController.show(application.id).absoluteURL(https, host),
-                s"Demande du ${Time.formatPatternFr(application.creationDate, datePattern)}"
-              )
-            )
-          )
-        ),
-        br,
-        br,
-        p(
-          "Comment avez-vous trouvé cet email ? ",
-          "Nous expérimentons. ",
-          "Pour nous aider, vous pouvez répondre à ce questionnaire : ",
-          a(
-            href := "https://startupdetat.typeform.com/to/PicUnQx4",
-            "https://startupdetat.typeform.com/to/PicUnQx4"
-          )
-        ),
-        br,
-        br,
-        p(
-          i(
-            "Si vous avez un problème ou besoin d’aide à propos de l’outil Administration+, contactez-nous sur ",
-            a(
-              href := s"mailto:${Constants.supportEmail}",
-              s"${Constants.supportEmail}"
-            )
-          )
-        )
-      )
+    val bodyInner = common.weeklyEmailBody(
+      infos,
+      application => routes.ApplicationController.show(application.id).absoluteURL(https, host),
+      daySinceLastAgentAnswerForApplicationsThatShouldBeClosed
+    )
     val email = Email(
-      subject = subject,
+      subject = common.weeklyEmailSubject,
       from = from,
       to = List(s"${quoteEmailPhrase(infos.user.name)} <${infos.user.email}>"),
-      bodyHtml = Some(div(bodyInner).toString)
+      bodyHtml = Some(common.renderEmail(bodyInner))
     )
     sendEmail(email)
   }
