+++ app/views/emails/WeeklyEmailInfos.scala
@@ -1,0 +1,5 @@
+package views.emails
+
+import models.{Application, User}
+
+case class WeeklyEmailInfos(user: User, applicationsThatShouldBeClosed: List[Application])
